+++ tests/client.js
@@ -1,5 +1,5 @@
 import { assert } from 'meteor/practicalmeteor:chai';
-import { createMeteorNetworkInterface } from 'meteor/apollo';
+import { createMeteorNetworkInterface, meteorClientConfig } from 'meteor/apollo';
 
 import ApolloClient from 'apollo-client';
 import gql from 'graphql-tag';
@@ -42,10 +42,101 @@ const personResult = {
 // Authenticate the test user
 Meteor._localStorage.setItem('Meteor.loginToken', 'foobar123');
 
-describe('Network interface', function() {
+describe('Meteor Client config', () => {
+  it('should accept a custom configuration object extending the default ones', () => {
+    const clientConfig = meteorClientConfig({
+      addTypename: false, // this is not a default option of the meteorClientConfig
+    });
+    
+    assert.isFalse(clientConfig.addTypename);
+  });
   
-  it('should create a network interface', function() {
-    assert.ok(createMeteorNetworkInterface({batchingInterface: false}));
+  it('should replace some options while keeping the other options', () => {
+    
+    const fakeCustomInterface = {
+      fakeProperty: 42,
+    };  
+    
+    // replace the default network interface to use by a fake one
+    // and replace the normalization function
+    const clientConfig = meteorClientConfig({
+      networkInterface: fakeCustomInterface,
+      dataIdFromObject: null,
+    });
+    
+    // we should still have access the ssrMode default configuration value
+    assert.deepEqual(clientConfig, {
+      networkInterface: {
+        fakeProperty: 42,
+      },
+      dataIdFromObject: null,
+      ssrMode: false,
+    });
+  });
+  
+  it('should extend the default config of the network interface', () => {
+    
+    // extend the opts ultimately passed to fetch
+    const networkInterface = createMeteorNetworkInterface({
+      opts: {
+        credentials: 'same-origin',
+      },
+    });  
+    
+    const clientConfig = meteorClientConfig({ networkInterface });
+    
+    // ApolloClient's 'createNetworkInterface' assign 'opts' to '_opts' in its constructor
+    assert.deepEqual(clientConfig.networkInterface._opts, { credentials: 'same-origin' });
+  });
+});
+
+describe('Network interface', () => {
+  
+  it('should create a network interface and not a batching interface', () => {
+    const networkInterface = createMeteorNetworkInterface({ batchingInterface: false });
+    
+    // as opposed to HTTPBatchedNetworkInterface
+    assert.equal(networkInterface.constructor.name, 'HTTPFetchNetworkInterface');
+  });
+  
+});
+
+describe('Query deduplication', () => {
+  
+  const randomQuery = gql`
+    query {
+      randomString
+    }`;
+  
+  it('does not deduplicate queries by default', () => {
+    
+    // we have two responses for identical queries, but only the first should be requested.
+    // the second one should never make it through to the network interface.
+    const client = new ApolloClient(meteorClientConfig());
+
+    const q1 = client.query({ query: randomQuery });
+    const q2 = client.query({ query: randomQuery });
+
+    // if deduplication happened, result2.data will equal data.
+    return Promise.all([q1, q2]).then(([result1, result2]) => {
+      assert.notDeepEqual(result1.data, result2.data);
+    });
+  });
+
+  it('deduplicates queries if the option is set', () => {
+
+    // we have two responses for identical queries, but only the first should be requested.
+    // the second one should never make it through to the network interface.
+    
+    const client = new ApolloClient(meteorClientConfig({queryDeduplication: true}));
+
+    const q1 = client.query({ query: randomQuery });
+    const q2 = client.query({ query: randomQuery });
+
+    // if deduplication didn't happen, result.data will equal data2.
+    return Promise.all([q1, q2]).then(([result1, result2]) => {
+      assert.deepEqual(result1.data, result2.data);
+    });
   });
   
 });